+++ FoggyWeather/FogType.cs
@@ -1,0 +1,10 @@
+﻿namespace FoggyWeather
+{
+    internal enum FogType
+    {
+        None = 0,
+        Normal = 1,
+        Blinding = 2,
+        Light = 3
+    }
+}