+++ HappyFishJump/JumpFish.cs
@@ -2,6 +2,8 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using StardewValley;
+using StardewValley.ItemTypeDefinitions;
+
 namespace HappyFishJump
 {
     public class JumpFish
@@ -59,18 +61,19 @@ namespace HappyFishJump
 
         public void Draw(SpriteBatch b)
         {
-            float angle = this.angle;
-            SpriteEffects effects = SpriteEffects.None;
+            float drawn_angle = this.angle;
+            SpriteEffects effect = SpriteEffects.None;
             if (this._flipped)
             {
-                effects = SpriteEffects.FlipHorizontally;
-                angle *= -1f;
+                effect = SpriteEffects.FlipHorizontally;
+                drawn_angle *= -1f;
             }
-            float num = 1f;
-            Vector2 globalPosition = this.position + new Vector2(0.0f, (float)Math.Sin((double)this._age / (double)this.jumpTime * Math.PI) * -this.jumpHeight);
+            float draw_scale = 1f;
+            Vector2 draw_position = this.position + new Vector2(0f, (float)Math.Sin((double)(this._age / this.jumpTime) * Math.PI) * (0f - this.jumpHeight));
             Vector2 origin = new(8f, 8f);
-            b.Draw(Game1.objectSpriteSheet, Game1.GlobalToLocal(Game1.viewport, globalPosition), new Rectangle?(Game1.getSourceRectForStandardTileSheet(Game1.objectSpriteSheet, this._fishObject.ParentSheetIndex, 16, 16)), Color.White, angle, origin, 4f * num, effects, 1f);
-            b.Draw(Game1.shadowTexture, Game1.GlobalToLocal(Game1.viewport, this.position), new Rectangle?(Game1.shadowTexture.Bounds), Color.White * 0.5f, 0.0f, new Vector2((float)(Game1.shadowTexture.Bounds.Width / 2), (float)(Game1.shadowTexture.Bounds.Height / 2)), 2f, effects, 1f);
+            ParsedItemData itemData = ItemRegistry.GetDataOrErrorItem(this._fishObject.QualifiedItemId);
+            b.Draw(itemData.GetTexture(), Game1.GlobalToLocal(Game1.viewport, draw_position), itemData.GetSourceRect(), Color.White, drawn_angle, origin, 4f * draw_scale, effect, this.position.Y / 10000f + 1E-06f);
+            b.Draw(Game1.shadowTexture, Game1.GlobalToLocal(Game1.viewport, this.position), Game1.shadowTexture.Bounds, Color.White * 0.5f, 0f, new Vector2(Game1.shadowTexture.Bounds.Width / 2, Game1.shadowTexture.Bounds.Height / 2), 2f, effect, this.position.Y / 10000f + 1E-06f);
         }
     }
 }