+++ StardewNotification/Configuration.cs
@@ -9,15 +9,19 @@
         public int RunNotificationsTime { get ; set; } = 0600; // 6:00 am
         public int BirthdayReminderTime { get; set; } = 1700; // 5:00 pm
         public bool NotifyFestivals { get; set; } = true;
+        public bool NotifyFestivalReminders { get; set; } = false;
         public bool NotifyBookseller { get; set; } = true;
         public bool NotifyTravelingMerchant { get; set; } = true;
         public bool NotifyToolUpgrade { get; set; } = false;
         public bool NotifyMaxLuck { get; set; } = true;
         public bool NotifyMinLuck { get; set; } = true;
+        public bool HideUnknownNPCBirthday { get; set; } = false;
         public bool NotifySeasonalForage { get; set; } = true;
         public bool ShowEmptyhay { get; set;} = true;
         public bool NotifyHay { get ; set; } = true;
         public bool ShowWeatherNextDay { get; set; } = true;
+
+        public bool IncludeGingerIsland { get; set; } = true;
         public int WeatherNextDayTime { get; set; } = 1700;  //5pm
         public bool NotifyTVChannels { get; set; } = true;
         public bool ShowSpringOnionCount { get; set; } = true;